elasticsearchapi/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
Django==5.2.6
djangorestframework==3.16.1
elastic-transport==9.1.0
elasticsearch==9.1.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
